feat(frontend): upgrade Satellite with CDN releases
